refactored src/lib/server/auth.ts into seperate files
