+++ src/lib/components/admin/EditUserModalForm.svelte
@@ -4,8 +4,9 @@
     import UserForm from "./UserForm.svelte"
     import HttpCodes from "$lib/utils/http-codes"
     import type { ResponseType } from "@api/user";
-    import type { EntireUser, InsertUserInput } from "$lib/server/database/auth"
+    import type { EntireUser, UpdateUserInput } from "$lib/server/database/auth/user";
     import type { PermissionGroup } from "@prisma/client";
+
 
     export let parent: SvelteComponent
 
@@ -28,7 +29,7 @@
         }
 	}
 
-    async function updateUserFormSubmit(userId: number, data: InsertUserInput): Promise<EntireUser | undefined> {
+    async function updateUserFormSubmit(userId: number, data: UpdateUserInput): Promise<EntireUser | undefined> {
         const response = await fetch(`/api/user/${user.id}`, {
             method: "PATCH",
             headers: {
