+++ src/lib/components/emails/EmailConfirmation.html
@@ -1,0 +1,16 @@
+<div style="display: flex; flex-direction: column; gap: 2rem;">
+    <div style="display: flex; flex-direction: row; align-items: center; gap: 0.5rem;">
+        <svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img" width="32" height="32" viewBox="0 0 36 36" class="iconify iconify--clarity"><path fill="currentColor" d="m33.48 29.63l-6.74-17.81a2 2 0 0 0-2.58-1.16L21 11.85V8.92A1.92 1.92 0 0 0 19.08 7H14V4.92A1.92 1.92 0 0 0 12.08 3H5a2 2 0 0 0-2 2v27a1 1 0 0 0 1 1h16a1 1 0 0 0 1-1V19.27l5 13.21a1 1 0 0 0 1.29.58l5.61-2.14a1 1 0 0 0 .58-1.29M12 8.83V31H5V5h7ZM19 31h-5V9h5Zm8.51-.25l-6.38-16.83l3.74-1.42l6.39 16.83Z" class="clr-i-outline clr-i-outline-path-1"></path><path fill="none" d="M0 0h36v36H0z"></path></svg>
+        <span style="font-weight: 700; font-size: 1.5rem;">Library</span>
+    </div>
+
+    <div style="display: flex; flex-direction: column; gap: 1rem;">
+        <h1 style="font-size: 1.5rem">Confirm your email address to get started</h1>
+        <p>Please click the link below to confirm your email address.</p>
+    </div>
+
+    <a href="{url}" style="width: fit-content; padding: 9px 1.25rem; display: inline-flex; align-items: center; justify-content: center; border-radius: 2px; gap: 0.5rem; background-color: rgb(234 134 26 / 0.2); box-shadow: inset 0 0 0 0px #fff, inset 0 0 0 calc(1px + 0px ) rgb( 234 134 26 / 1 ), 0 0 #0000">
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" class="feather feather-external-link"><path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path><polyline points="15 3 21 3 21 9"></polyline><line x1="10" y1="14" x2="21" y2="3"></line></svg>
+        <span>Confirm Email</span>
+    </a>
+</div>
