+++ src/lib/server/database/auth/index.ts
@@ -1,0 +1,14 @@
+import user from "./user"
+import session from "./session"
+import emailConfirmation from "./email-confirmation"
+import readingState from "./reading-state"
+import permissionGroup from "./permission-group"
+
+
+export default {
+    user,
+    session,
+    emailConfirmation,
+    readingState,
+    permissionGroup
+}
