+++ src/lib/utils/generate-expiration-date.ts
@@ -1,0 +1,8 @@
+export function generateExpirationDate(time: number): Date {
+    const now = new Date()
+    now.setSeconds(now.getSeconds() + time)
+    return now
+}
+
+
+export default generateExpirationDate
