+++ src/lib/utils/is-date-expired.ts
@@ -1,0 +1,6 @@
+export function isDateExpired(date: Date): boolean {
+    return date.getTime() < Date.now()
+}
+
+
+export default isDateExpired
