+++ src/routes/admin/initial-setup/+page.server.ts
@@ -6,20 +6,20 @@ import { ADMIN_EMAIL } from "$env/static/private"
 
 
 export const load: PageServerLoad = async () => {
-    const usersCount = await db.auth.getUsersCount()
+    const usersCount = await db.auth.user.getUserCount()
 
     if (usersCount > 0) {
         redirect(HttpCodes.SeeOther, "/admin")
     }
 
     if (ADMIN_EMAIL) {
-        const user = await db.auth.createUser({
+        const user = await db.auth.user.createUser({
             email: ADMIN_EMAIL,
             username: "Admin",
             permissionGroup: "Admin"
         })
 
-        const err = await db.auth.sendConfirmationEmail(user, "/admin")
+        const err = await db.auth.emailConfirmation.sendConfirmationEmailAndSaveRequest(user, "/admin")
         if (err) error(HttpCodes.ServerError.InternalServerError, {
             error: undefined,
             message: "Error sending confirmation email"
