+++ src/routes/auth/sign-out/+server.ts
@@ -9,7 +9,7 @@ export const GET: RequestHandler = async ({ cookies, url }) => {
     const sessionToken = cookies.get("sessionToken")
 
     if (sessionToken && validate(sessionToken)) {
-        await db.auth.deleteSessionByToken(sessionToken)
+        await db.auth.session.deleteSessionByToken(sessionToken)
         cookies.delete("sessionToken", {
             path: "/"
         })
